update pathfinder trie
